+++ Nodes/Gates/ButtonNode.cs
@@ -18,7 +18,7 @@ namespace Wired.Nodes
             base.Awake();
         }
 
-        public override void IncreaseVoltage(uint amount)
+        public override void IncreaseVoltage(float amount)
         {
             if (!allowCurrent)
                 return;
@@ -28,7 +28,7 @@ namespace Wired.Nodes
         }
 
 
-        public override void DecreaseVoltage(uint amount)
+        public override void DecreaseVoltage(float amount)
         {
             if (Voltage < amount)
                 Voltage = 0;
